[수정] 별점 기능 수정
